# Optimise reset connection