--- a/src/MySqlConnector/Protocol/Serialization/StandardPayloadHandler.cs
+++ b/src/MySqlConnector/Protocol/Serialization/StandardPayloadHandler.cs
@@ -23,6 +23,8 @@ namespace MySqlConnector.Protocol.Serialization
 			m_sequenceNumber = 0;
 		}
 
+		public void SetNextSequenceNumber(int sequenceNumber) => m_sequenceNumber = sequenceNumber;
+
 		public IByteHandler ByteHandler
 		{
 			get => m_byteHandler;