@bgrainger 2018-04-18
Fixed in [0.40.0](https://github.com/mysql-net/MySqlConnector/releases/tag/0.40.0).